add orcafs snapshot api

+++ server/service/snapshot.js
@@ -1,4 +1,5 @@
 const config = require('../config');
+const afterMe = require('./afterMe');
 const database = require('./database');
 const socket = require('../module/socket');
 const promise = require('../module/promise');
@@ -17,7 +18,11 @@ const model = {
     },
     async updateSnapshotSetting(param) {
         let { total, manual, auto } = param;
-        await database.updateSetting({ key: 'snapshotsetting' }, { value: { total, manual, auto } });
+        let { errorId, message } = await afterMe.updateSnapshotSetting({ total, manual, schedule: auto });
+        if (!errorId) {
+            await database.updateSetting({ key: 'snapshotsetting' }, { value: { total, manual, auto } });
+        }
+        return { errorId, message };
     },
     async getSnapshot(param) {
         return await database.getSnapshot(param);
@@ -26,6 +31,7 @@ const model = {
         return Boolean(await database.getSnapshotCount({ creating: true }) + await database.getSnapshotCount({ deleting: true }) + await database.getSnapshotCount({ rollbacking: true }));
     },
     async createSnapshot(param) {
+        let result = false;
         let { name, description, isAuto = false, creating = true, deleting = false, rollbacking = false, createTime = new Date() } = param;
         let setting = await model.getSnapshotSetting();
         let limit = Number(setting.manual);
@@ -33,40 +39,82 @@ const model = {
         if (count < limit) {
             await database.addSnapshot({ name, description, isAuto, creating, deleting, rollbacking, createTime });
             socket.postEventStatus('snapshot', 11, name, true, false);
-            await promise.runTimeOutInPromise(10);
-            await database.updateSnapshot({ name }, { creating: false });
-            socket.postEventStatus('snapshot', 12, name, true, true);
-            return true;
+            let { errorId, data, message } = await afterMe.createSnapshot({ name, schedule: String(isAuto) });
+            if (!errorId) {
+                await database.updateSnapshot({ name }, { createTime: new Date(data.createTime), creating: false });
+                socket.postEventStatus('snapshot', 12, name, true, true);
+                result = true;
+            } else {
+                handler.error(132, message, param);
+                await database.deleteSnapshot({ name });
+                socket.postEventStatus('snapshot', 12, name, false, true);
+                result = false;
+            }
         } else {
             socket.postEventStatus('snapshot', 12, name, false, true);
-            return false;
+            result = false;
         }
+        return result;
     },
     async updateSnapshot(param) {
         let { name, description } = param;
         await database.updateSnapshot({ name }, { description });
     },
     async deleteSnapshot(param) {
+        let result = false;
         let { name } = param;
         await database.updateSnapshot({ name }, { deleting: true });
-        await promise.runTimeOutInPromise(5);
-        await database.deleteSnapshot({ name });
+        let { errorId, message } = await afterMe.deleteSnapshot({ name });
+        if (!errorId) {
+            await database.deleteSnapshot({ name });
+            socket.postEventStatus('snapshot', 13, name, true, true);
+            result = true;
+        } else {
+            handler.error(134, message, param);
+            await database.updateSnapshot({ name }, { deleting: false });
+            socket.postEventStatus('snapshot', 14, name, false, true);
+            result = false;
+        }
+        return result;
     },
     async batchDeleteSnapshot(param) {
+        let result = false;
         let { names } = param;
         for (let name of names) {
             await database.updateSnapshot({ name }, { deleting: true });
         }
-        await promise.runTimeOutInPromise(5);
-        for (let name of names) {
-            await database.deleteSnapshot({ name });
+        let { errorId } = await afterMe.batchDeleteSnapshot({ names });
+        if (!errorId) {
+            for (let name of names) {
+                await database.deleteSnapshot({ name });
+            }
+            socket.postEventStatus('snapshot', 15, { total: names.length }, true, true);
+            result = true;
+        } else {
+            handler.error(135, message, param);
+            for (let name of names) {
+                await database.updateSnapshot({ name }, { deleting: false });
+            }
+            socket.postEventStatus('snapshot', 16, { total: names.length }, false, true);
+            result = false;
         }
+        return result;
     },
     async rollbackSnapshot(param) {
+        let result = false;
         let { name } = param;
         await database.updateSnapshot({ name }, { rollbacking: true });
-        await promise.runTimeOutInPromise(20);
+        let { errorId } = await afterMe.rollbackSnapshot({ name });
         await database.updateSnapshot({ name }, { rollbacking: false });
+        if (!errorId) {
+            socket.postEventStatus('snapshot', 18, name, true, true);
+            result = true;
+        } else {
+            handler.error(136, message, param);
+            socket.postEventStatus('snapshot', 18, name, false, true);
+            result = false;
+        }
+        return result;
     },
     async getSnapshotSchedule(param) {
         return await database.getSnapshotSchedule(param);
